pyproject.toml
[project]
name = "discordchatimporter"
version = "0.1.0"
description = "A Discord bot for importing chat messages from various sources"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.0.0",
    "disnake>=2.9.0",
    "pyimgur>=0.5.0",
    "emoji>=2.0.0",
    "python-dotenv>=1.0.0",
    "beautifulsoup4>=4.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-sugar>=1.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.8.0",
    "basedpyright>=1.21.0",
    "codespell>=2.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core", "cogs"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-sugar>=1.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.8.0",
    "basedpyright>=1.21.0",
    "codespell>=2.3.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100
extend-exclude = ["format_json/"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "C901",   # too complex
]

[tool.ruff.lint.isort]
known-first-party = ["discordchatimporter"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=core --cov=cogs --cov-report=term-missing"

[tool.basedpyright]
include = ["core", "cogs", "main.py"]
exclude = ["format_json", "tests/**/test_*.py"]
pythonVersion = "3.11"
pythonPlatform = "All"
typeCheckingMode = "standard"
reportMissingTypeStubs = false
reportUnknownMemberType = false
